% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaps.R
\name{scaps}
\alias{scaps}
\title{Turning strings into small capitals}
\usage{
scaps(x)
}
\arguments{
\item{x}{A character string (e.g. a word or phrase).}
}
\value{
Character string with HTML-tag for small capitals.
}
\description{
This is a wrapper to put HTML-tag around character string input to be rendered as small capital in the output document of R Markdown file.
}
\examples{
scaps("happiness")
}
