% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaps_bold.R
\name{scaps_bold}
\alias{scaps_bold}
\title{Turning strings into boldfaced small-capitals}
\usage{
scaps_bold(x)
}
\arguments{
\item{x}{A character string (e.g. a word or phrase).}
}
\value{
Boldfaced, small capital character string.
}
\description{
This is a wrapper to put HTML-tag around character string input to be rendered as \strong{boldfaced} small capital in the output document of an R Markdown file.
}
\examples{
scaps_bold("happiness")
}
