% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaps_boldfaceitalics.R
\name{scaps_bolditalics}
\alias{scaps_bolditalics}
\title{Turning strings into boldfaced, italicised small-capitals}
\usage{
scaps_bolditalics(x)
}
\arguments{
\item{x}{A character string (e.g. a word or phrase).}
}
\value{
Boldfaced as well as italicised small capital character string.
}
\description{
This is a wrapper to put HTML-tag around character string input to be rendered as \strong{boldfaced} as well as \emph{italicised} small capital in the output document of an R Markdown file.
}
\examples{
\dontrun{
scaps_bolditalics("happiness")
}
}
